backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
